+++ .images.yaml
@@ -1,46 +1,45 @@
-mirror:
-    registry: plexdev.azurecr.io
-    repository: pcp
+target:
+    registry: lol.com
 images:
-  - repository: nginx
+  - repository: pcp/nginx
     version: 1.14.2
-    origin: docker.io
-  - repository: openpolicyagent/gatekeeper
+    source: docker.io
+  - repository: pcp/openpolicyagent/gatekeeper
     version: v3.1.0-beta.10
-    origin: docker.io
-  - repository: kubernetes-ingress-controller/nginx-ingress-controller
+    source: docker.io
+  - repository: pcp/kubernetes-ingress-controller/nginx-ingress-controller
     version: 0.32.0
-    origin: quay.io
-  - repository: coreos/kube-state-metrics
+    source: quay.io
+  - repository: pcp/coreos/kube-state-metrics
     version: v1.9.5
-    origin: quay.io
-  - repository: coreos/prometheus-operator
+    source: quay.io
+  - repository: pcp/coreos/prometheus-operator
     version: v0.40.0
-    origin: quay.io
-  - repository: jimmidyson/configmap-reload
+    source: quay.io
+  - repository: pcp/jimmidyson/configmap-reload
     version: v0.3.0
-    origin: docker.io
-  - repository: coreos/prometheus-config-reloader
+    source: docker.io
+  - repository: pcp/coreos/prometheus-config-reloader
     version: v0.40.0
-    origin: quay.io
-  - repository: consul
+    source: quay.io
+  - repository: pcp/consul
     version: 1.7.2
-    origin: docker.io
-  - repository: prom/node-exporter
+    source: docker.io
+  - repository: pcp/prom/node-exporter
     version: v0.18.1
-    origin: docker.io
-  - repository: grafana/promtail
+    source: docker.io
+  - repository: pcp/grafana/promtail
     version: v0.4.0
-    origin: docker.io
-  - repository: twistlock/defender
+    source: docker.io
+  - repository: pcp/twistlock/defender
     version: defender_20_04_177
-    origin: registry.twistlock.com
-  - repository: jettech/kube-webhook-certgen
+    source: registry.twistlock.com
+  - repository: pcp/jettech/kube-webhook-certgen
     version: v1.2.0
-    origin: docker.io
-  - repository: prom/alertmanager
+    source: docker.io
+  - repository: pcp/prom/alertmanager
     version: v0.21.0
-    origin: docker.io
-  - repository: prom/prometheus
+    source: docker.io
+  - repository: pcp/prom/prometheus
     version: v2.19.2
-    origin: docker.io
+    source: docker.io
